Move field test base into propagator class
